+++ src/validate_submission.py
@@ -24,15 +24,16 @@ def make_parser():
 Validate a HuBMAP submission, both the metadata TSVs, and the datasets,
 either local or remote, or a combination of the two.''',
         epilog='''
-Typical usecases:
-  --tsv_paths: Used to validate TSVs in isolation, without checking references.
+Typical usage:
+  --tsv_paths: Used to validate Sample metadata TSVs. (Because it does
+  not check references, should not be used to validate Dataset metadata TSVs.)
 
-  --local_directory: Used in development against test fixtures, and could be used
-  by labs before submission.
+  --local_directory: Used by lab before submission, and on Globus after upload.
 
   --local_directory + --dataset_ignore_globs + --submission_ignore_globs:
-  Currently, during ingest, the metadata TSVs are broken up, and one-line TSVs
-  are put in each dataset directory. This structure needs extra ignores.
+  After the initial validation on Globus, the metadata TSVs are broken up,
+  and one-line TSVs are put in each dataset directory. This structure needs
+  extra parameters.
 
 ''',
         formatter_class=argparse.RawDescriptionHelpFormatter)